+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Calendar Versioning](https://calver.org).
 
 ### Added
 - [CSE] Added support for the [credential] ManagementObject specialization.
+- [CSE] Added support for other data types (int, float, bool, list, dict) for the &lt;ContentInstance>'s *content* attribute.
 - [TUI] Added support for enablung/disabling syntax highlighting in the resource editor of the text UI. The syntax package is now not installed by default because it is not available on all platforms. It be installed afterward if needed.
 - [TESTS] Added cmdline argument to disable the check and use of the Upper Tester interface.